Revision summary: -Changes in logging system with the ability to rollback to the original console-logging. -logWrite macro changed to log(format,va_args) and logEx(field,format,va_args) -*.hpp <- *.h , \\ <- / -Removed warnings -Type conversion fixes. There's still a lot of work to be done here..Especially in template handling which could affect generated code(if not already.. which is what i believe).
Revision notes (programmers):
To rollback to original logging method that you're used to , open log/logging.h
and replace "#undef DEFAULT_LOGGING" with "#define DEFAULT_LOGGING".

Revision notes (end users):
For maximum performance it is recommended to leave logging disabled.
If you encounter a bug , enable it to provide the call trace.

+++ nulldc/nullDC/dc/sh4/rec_v1/blockmanager.cpp
@@ -601,7 +601,7 @@ CompiledBlockInfo* bm_ReverseLookup(void* code_ptr)
 	{
 		CompiledBlockInfo* b=all_block_list[i];
 
-		if ((c-(u8*)b->Code) < b->size)
+		if ( (u32)(c-(u8*)b->Code) < b->size)
 		{
 			return b;
 		}
@@ -614,7 +614,7 @@ CompiledBlockInfo* bm_ReverseLookup(void* code_ptr)
 	{
 		CompiledBlockInfo* b=SuspendedBlocks[i];
 
-		if ((c-(u8*)b->Code) < b->size)
+		if ( (u32)(c-(u8*)b->Code) < b->size)
 		{
 			return b;
 		}
